+++ starkbank/__init__.py
@@ -107,3 +107,6 @@ from .split.__split import Split
 
 from . import splitreceiver
 from .splitreceiver.__splitreceiver import SplitReceiver
+
+from . import splitprofile
+from .splitprofile.__splitprofile import SplitProfile