+++ starkbank/splitprofile/__init__.py
@@ -1,0 +1,3 @@
+from .__splitprofile import put, query, get, page
+from .log.__log import Log
+from . import log